@@ -7,17 +7,16 @@ version: 0.2.0+3
 homepage: https://github.com/james-alex/flutter_palette
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_color_models: ^0.3.3+1
-  meta: ^1.1.8
-  palette: ^0.2.0+1
-  unique_list: ^0.1.2+1
+  flutter_color_models: ^1.1.0+1
+  palette: ^1.0.0
+  unique_list: ^1.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.0
+  pedantic: ^1.11.0
