@@ -6,5 +6,5 @@ library flutter_palette;
 export 'package:flutter/painting.dart' show Color;
 export 'package:flutter_color_models/flutter_color_models.dart';
 export 'package:palette/palette.dart'
-    show ColorSpace, ColorSortingProperty, ColorSortingDirection;
+    show ColorSortingProperty, ColorSortingDirection;
 export 'src/color_palette.dart';
